# openapi.yaml
openapi: 3.0.0
info:
  title: Car CRUD and authorizations
  version: 0.1.0
servers:
- description: Mock Server
  url: http://localhost:3000
...
paths:
...
  /api/v1/get-user/{id}:
    get:
      operationId: getUser
      description: Enpoint get user bekerja dengan memberikan semua data user baik admin maupun member dari api ini. Endpoint getUser hanya bisa diakses oleh pemilik akses level super admin (level 2)
      tags:
      - Super Admin Authorization
      parameters:
      - name: x-access-token
        in: header
        required: true
        schema:
          type: string
          format: jwt
          description: Parameter token dengan nama x-access-token adalah parameter wajib untuk mengakses endpoint enpoint pada API ini. x-access-token, bisa didaptakan dari response endpoint login dengan key token
          example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiNjY1M2YzNjktNGRmNC00M2Y5LWE2NzMtOWVjNjEwMzU5NDI1IiwiaWF0IjoxNjY2MDcwNjA3fQ.a2uU6WZ_wW6zfZOTSYzSLusiuTTczLYCvW7GBfioMvo
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Parameter id ini, merupakan id berformat uuid dari user yang ditargetkan.
          example: ee831b29-47e3-495f-82b9-e45797044d89
      responses:
        "200":
          description: Berikut adalah bentuk response success dari endpoint getUser
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userData"
        "400":
          description: Berikut bentuk response kesalahan server / gagal memanggil endpoint getUser
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/badRequest"
        "401":
          description: Hasil respon berikut adalah contoh pelanggaran hak akses oleh user yang tidak mempunyai akses level 2 / super admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/unauthorized"
...
components:
  schemas:
    carType:
      type: object
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: MPV
        createdAt:
          type: string
          example: '2022-10-15T04:19:51.834Z'
        updatedAt:
          type: string
          example: '2022-10-15T04:19:51.834Z'
    carData:
      type: object
      properties:
        id:
          type: string
          example: 8a90d142-6b29-498c-938f-f862203093b7
        name:
          type: string
          example: Carrera 2765
        type_car:
          type: integer
          example: 2
        price:
          type: integer
          example: 100000
        available:
          type: boolean
          example: true
        available_at:
          type: string
          example: '2022-03-23T15:49:05.563Z'
        deleted:
          type: boolean
          example: false
        image:
          type: string
          example: /images/ee831b29-47e3-495f-82b9-e45797044d89_image_1666015992628_768365436.jpg
        user_id:
          type: string
          example: ee831b29-47e3-495f-82b9-e45797044d89
        capacity:
          type: integer
          example: 6
        description:
          type: string
          example: lorem ipsum dolor sit amet
        transmission:
          type: string
          example: manual
        year:
          type: integer
          example: 2012
        size:
          type: string
          example: medium
        createdAt:
          type: string
          example: '2022-10-17T14:13:12.647Z'
        updatedAt:
          type: string
          example: '2022-10-17T14:13:12.647Z'
        type:
          $ref: "#/components/schemas/carType"
...
    userData:
      type: object
      properties:
        id:
          type: string
          example: ee831b29-47e3-495f-82b9-e45797044d89
        username:
          type: string
          example: admin
        first_name:
          type: string
          example: admin
        last_name:
          type: string
          example: admin
        email:
          type: string
          example: admin@mail.com
        password:
          type: string
          example: $2a$10$X7V.FfeIM09h57xISUivxuyaQ3UmjXQcUL.Y6QcgL3icAar5PF4jm
        access_level:
          type: integer
          example: 1
        createdAt:
          type: string
          example: 2022-10-15T04:19:51.853Z
        updatedAt:
          type: string
          example: 2022-10-18T10:23:07.277Z
        cars:
          type: array
          items:
            $ref: "#/components/schemas/carData"
...
    badRequest:
      type: object
      properties:
        status:
          type: string
          example: FAIL
        message:
          type: string
          example: (server error ...)
...
    unauthorized:
      type: object
      properties:
        error:
          type: string
          example: Authentication rejected. You not an ...
...